requirements.txt
# YOLO11 Training on CrowdHuman Dataset
# Install with: pip install -r requirements.txt

# Core YOLO library
ultralytics>=8.3.0

# Dataset handling
datasets>=2.14.0
huggingface_hub>=0.19.0

# Image processing
opencv-python>=4.8.0
Pillow>=10.0.0

# Utilities
tqdm>=4.66.0
numpy>=1.24.0

# Optional: For faster training
# torch>=2.0.0  # Usually installed with ultralytics

